Update patch set 2
Patch Set 2: Code-Review-1

(3 comments)

Please update the module version number.  Also, please set the default syslog facility to an empty string, as the default of "LOG_USER" will not work, even though, internally, that is the default selected when the facility is not specified for the audisp syslog plugin.

Patch-set: 2
Label: Code-Review=-1