(SIMP-2558) Changed the default failure to printk
The default failure state for auditd was to panic the system. However,
with the required audit rules for chmod and chown, this was routinely
overflowing the large log buffer that we had put in place.

Instead of arbitrarily increasing the log buffer, we opted to change
the failure state to logging printk messages until we can figure out
how to properly eliminate this failure case.

SIMP-2558 #close