(SIMP-7838) Allow pre-configured profiles
This change allows the use of included sample rulesets for EL8/audit 3.x
to configure the system. Adds a fact to list the availble rulesets, adds
the ability to skip configuration of SIMP audit profiles, and adds configuration
class for moving sample rulesets into place. This also provides a mechanism to
generate the privileged ruleset from the embedded set of commands in the sample.

SIMP-7838 #close