(SIMP-7838) Allow pre-configured profiles (#118)
This change allows the use of included sample rulesets for EL8/audit 3.x
to configure the system via a built-in profile. Adds a fact to list the availble rulesets, adds
the ability to skip configuration of SIMP audit profiles, and adds configuration
class for moving sample rulesets into place. This also provides a mechanism to
generate the privileged ruleset from the embedded set of commands in the sample.

Allows the user to disable all 'SIMP'-standard functionality by ONLY
including the `built_in` profile and adds tests for enabling the built-in
functionality alongside another existing profile set.

Allow setting memory from the gitlab-ci yaml file, ERB's the setting
in the nodesets.

SIMP-7838 #close